+++ Tests/SentryTests/Protocol/SentryEventTests.swift
@@ -85,4 +85,13 @@ class SentryEventTests: XCTestCase {
         let event = Event(error: error)
         XCTAssertEqual(error, event.error as? CocoaError)
     }
+
+    func testSerializeWithoutMessage() {
+        let actual = Event().serialize()
+        XCTAssertNil(actual["message"])
+    }
+
+    func testMessageIsNil() {
+        XCTAssertNil(Event().message)
+    }
 }