+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - feat: Add error to SentryEvent #944
+- fix: Mark SentryEvent.message as Nullable #943
 - fix: Stacktrace inApp marking on Simulators #942
 - feat: Group NSError by domain and code #941
 - fix: Discard Sessions when JSON is faulty #939